Added code

+++ Homeworks_Projects/BurgerApp/BurgerApp/BurgerApp/Models/Domain/Burger.cs
@@ -4,12 +4,12 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public bool IsVegeterian { get; set; }
         public bool IsVegan { get; set; }
         public bool HasFries { get; set; }
 
-        public Burger(int id, string name, float price, bool isVegetarian, bool isVegan, bool hasFries)
+        public Burger(int id, string name, double price, bool isVegetarian, bool isVegan, bool hasFries)
         {
             Id = id;
             Name = name;
